+++ geotrek/tourism/models.py
@@ -286,9 +286,9 @@ class TouristicContent(AddPropertyMixin, PublishableMixin, MapEntityMixin, Struc
     type2 = models.ManyToManyField(TouristicContentType, related_name='contents2',
                                    verbose_name=_(u"Type 2"), db_table="t_r_contenu_touristique_type2",
                                    blank=True)
-    source = models.ForeignKey('common.RecordSource',
-                               null=True, blank=True, related_name='touristiccontents',
-                               verbose_name=_("Source"), db_column='source')
+    source = models.ManyToManyField('common.RecordSource',
+                                    null=True, blank=True, related_name='touristiccontents',
+                                    verbose_name=_("Source"), db_table='t_r_contenu_touristique_source')
     eid = models.CharField(verbose_name=_(u"External id"), max_length=128, blank=True, db_column='id_externe')
 
     objects = NoDeleteMixin.get_manager_cls(models.GeoManager)()
@@ -381,9 +381,9 @@ class TouristicEvent(AddPropertyMixin, PublishableMixin, MapEntityMixin, Structu
     target_audience = models.CharField(verbose_name=_(u"Target audience"), max_length=128, blank=True, null=True, db_column='public_vise')
     practical_info = models.TextField(verbose_name=_(u"Practical info"), blank=True, db_column='infos_pratiques',
                                       help_text=_(u"Recommandations / To plan / Advices"))
-    source = models.ForeignKey('common.RecordSource',
-                               null=True, blank=True, related_name='touristicevents',
-                               verbose_name=_("Source"), db_column='source')
+    source = models.ManyToManyField('common.RecordSource',
+                                    null=True, blank=True, related_name='touristicevents',
+                                    verbose_name=_("Source"), db_table='t_r_evenement_touristique_source')
     eid = models.CharField(verbose_name=_(u"External id"), max_length=128, blank=True, db_column='id_externe')
 
     objects = NoDeleteMixin.get_manager_cls(models.GeoManager)()