+++ geotrek/trekking/forms.py
@@ -133,7 +133,7 @@ class TrekForm(BaseTrekForm):
 
         # Since we use chosen() in trek_form.html, we don't need the default help text
         for f in ['themes', 'networks', 'accessibilities',
-                  'web_links', 'information_desks']:
+                  'web_links', 'information_desks', 'source']:
             self.fields[f].help_text = ''
 
     def clean_duration(self):