+++ geotrek/trekking/models.py
@@ -82,9 +82,9 @@ class Trek(StructureRelated, PicturesMixin, PublishableMixin, MapEntityMixin, To
                                                help_text=_(u"Where to obtain information"))
     points_reference = models.MultiPointField(verbose_name=_(u"Points of reference"), db_column='geom_points_reference',
                                               srid=settings.SRID, spatial_index=False, blank=True, null=True)
-    source = models.ForeignKey('common.RecordSource',
-                               null=True, blank=True, related_name='treks',
-                               verbose_name=_("Source"), db_column='source')
+    source = models.ManyToManyField('common.RecordSource',
+                                    null=True, blank=True, related_name='treks',
+                                    verbose_name=_("Source"), db_table='o_r_itineraire_source')
     eid = models.CharField(verbose_name=_(u"External id"), max_length=128, blank=True, db_column='id_externe')
 
     objects = Topology.get_manager_cls(models.GeoManager)()